+++ csharp-client/SupermarketSimArchipelago/Patches/EconomyPatches.cs
@@ -1,4 +1,5 @@
 using HarmonyLib;
+using SupermarketSimArchipelago;
 using UnityEngine;
 
 namespace SupermarketArchipelago
@@ -20,9 +21,14 @@ namespace SupermarketArchipelago
         }
 
         [HarmonyPostfix]
-        public static void Postfix()
+        public static void Postfix(MoneyManager.TransitionType type)
         {
             ArchipelagoMoneyHandler.CheckMoneyMilestones();
+
+            if (type == MoneyManager.TransitionType.CHECKOUT_INCOME)
+            {
+                ArchipelagoCheckoutHandler.CheckCustomerCheckoutLocation();
+            }
         }
     }
 
@@ -50,37 +56,7 @@ namespace SupermarketArchipelago
         }
     }
 
-    // ==========================================
-    // ARCHIPELAGO MONEY MILESTONE HANDLER
-    // ==========================================
-    public static class ArchipelagoMoneyHandler
-    {
-        public static void CheckMoneyMilestones()
-        {
-            if (!ArchipelagoConfig.EnableMoneyMilestones) return;
-            if (!ArchipelagoClient.IsConnected) return;
-            if (MoneyManager.Instance == null) return;
-           
-
-            float currentMoney = MoneyManager.Instance.Money;
-
-            int maxMoney = ArchipelagoConfig.MaxMoneyMilestone;
-            int interval = ArchipelagoConfig.MoneyMilestoneInterval;
-
-            for (int money = interval; money <= maxMoney; money += interval)
-            {
-                if (currentMoney >= money)
-                {
-                    long locationId = ArchipelagoIdHelper.FromMoneyMilestone(money);
-                    if (!ArchipelagoClient.CheckLocationAlreadySent(locationId))
-                        ArchipelagoClient.SendLocation(locationId);
-
-                }
-            }
-        }
-    }
-
-
+   
 
 
 }