+++ csharp-client/SupermarketSimArchipelago/Patches/StoragePatches.cs
@@ -1,4 +1,4 @@
-﻿using HarmonyLib;
+using HarmonyLib;
 using UnityEngine;
 using UnityEngine.Localization.Components;
 using SupermarketSimArchipelago;